chore: feature guard
